% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{country_cases_cumulative}
\alias{country_cases_cumulative}
\title{country_cases_cumulative}
\usage{
country_cases_cumulative(cv_cases, start_point, plot_start_date)
}
\arguments{
\item{cv_cases}{dataframe}

\item{start_point}{character string indicating point of origin}

\item{plot_start_date}{Date variable}
}
\description{
a function to plot cumulative cases of covid by region
}
