% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{cumulative_plot}
\alias{cumulative_plot}
\title{cumulatve_plot}
\usage{
cumulative_plot(cv_aggregated, plot_date)
}
\arguments{
\item{cv_aggregated}{dataframe}

\item{plot_date}{Date variable}
}
\description{
a function to plot cumulative cases of covid
}
