% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cv_aggregated}
\alias{cv_aggregated}
\title{cv_aggregated}
\format{
a dataframe with 108 rows and 4 columns
}
\usage{
cv_aggregated
}
\description{
Dataset of aggregated COVID-19 cases
}
\keyword{datasets}
