% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cv_cases_continent}
\alias{cv_cases_continent}
\title{cv_cases_continent}
\format{
a dataframe with 646 rows and 14 columns
}
\usage{
cv_cases_continent
}
\description{
Dataset of COVID-19 cases at continental level
}
\keyword{datasets}
