% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cv_cases_global}
\alias{cv_cases_global}
\title{cv_cases_global}
\format{
a dataframe with 108 rows and 13 columns
}
\usage{
cv_cases_global
}
\description{
Dataset of COVID-19 cases at global level
}
\keyword{datasets}
