% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cv_large_countries}
\alias{cv_large_countries}
\title{cv_large_countries}
\format{
a dataframe with 193 rows and 29 columns
}
\usage{
cv_large_countries
}
\description{
Dataset of  COVID-19 cases
}
\keyword{datasets}
