% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cv_today_reduced}
\alias{cv_today_reduced}
\title{cv_today_reduced}
\format{
a dataframe with 188 rows and 29 columns
}
\usage{
cv_today_reduced
}
\description{
Dataset of  COVID-19 cases
}
\keyword{datasets}
