% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about_server}
\alias{mod_about_server}
\title{mod_about_server}
\usage{
mod_about_server(id)
}
\arguments{
\item{id}{internal shiny function}
}
\description{
mod_about_server() function
}
\keyword{internal}
