% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prediction.R
\name{mod_prediction_server}
\alias{mod_prediction_server}
\title{prediction server function}
\usage{
mod_prediction_server(id, cv_states)
}
\arguments{
\item{id}{internal parameter for shiny}

\item{cv_states}{dataset containing covid19 records}
}
\value{
a shiny module server
}
\description{
the server logic for the prediction module
}
