% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prediction.R
\name{mod_prediction_ui}
\alias{mod_prediction_ui}
\title{prediction ui function}
\usage{
mod_prediction_ui(id)
}
\arguments{
\item{id}{internal parameter for shiny}
}
\value{
a shiny ui taglist
}
\description{
a shiny module to compare xgboost and sir models across all states
}
