% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{new_cases_plot}
\alias{new_cases_plot}
\title{new_cases_plot}
\usage{
new_cases_plot(cv_aggregated, plot_date)
}
\arguments{
\item{cv_aggregated}{dataframe}

\item{plot_date}{Date variable}
}
\description{
a function to plot new cases of covid
}
