% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{pal}
\alias{pal}
\title{create a color palette using leaflet::colorBin}
\usage{
pal(bins, df)
}
\arguments{
\item{bins}{numeric vector}

\item{df}{dataframe}
}
\description{
create a color palette using leaflet::colorBin
}
