% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{worldcountry}
\alias{worldcountry}
\title{worldcountry}
\format{
sp of 241 countries
}
\usage{
worldcountry
}
\description{
Dataset of world country polygons as a shape file
}
\keyword{datasets}
