% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about_ui}
\alias{mod_about_ui}
\title{mode_about_ui and mod_about_server}
\usage{
mod_about_ui(id)
}
\arguments{
\item{id}{Internal parameters for \code{shiny}.}
}
\description{
mod_about_ui() function
}
\keyword{internal}
